Alternate Extruder Motor Mount for M2
This one can be made out of metal (milled) or printed. It holds the
motor square event if it is made of plastic and the motor gets warm.
You might want to use a pair of Heat Set Inserts for securing the top
bridge member if using plastic or use self tapping screws.